organizing file structure, deleting unused files
