+++ src/components/Landing/screenshot.js
@@ -4,7 +4,7 @@ import { Gallery, GalleryImage } from 'react-gesture-gallery';
 const images = [
   'https://i.ibb.co/4sPC32f/image.png',
   'https://i.ibb.co/Dwj7BjP/image-1.png',
-  'https://i.ibb.co/k9659RJ/image-6.png'
+  'https://i.ibb.co/k9659RJ/image-6.png',
 ];
 
 const GalleryImg = React.memo(GalleryImage);
@@ -17,7 +17,7 @@ function Screenshot() {
       if (index === 2) {
         setIndex(0);
       } else {
-        setIndex(prev => prev + 1);
+        setIndex((prev) => prev + 1);
       }
     }, 3000);
     return () => clearInterval(timer);
@@ -29,20 +29,18 @@ function Screenshot() {
         background: 'white',
         height: '60vh',
         width: '100vw',
-        transition: "opacity 0.5s ease",
-        transform: "translateX(-50%)",
+        transition: 'opacity 0.5s ease',
+        transform: 'translateX(-50%)',
         display: 'flex',
         flexDirection: 'row',
         marginLeft: '50%',
-
-
       }}
       index={index}
-      onRequestChange={i => {
+      onRequestChange={(i) => {
         setIndex(i);
       }}
     >
-      {images.map(image => (
+      {images.map((image) => (
         <GalleryImg objectFit="contain" key={image} src={image} />
       ))}
     </Gallery>
