+++ src/pages/App/contacts.js
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
-import { GET_USER_CONNECTIONS } from '../queries/index';
+import { GET_USER_CONNECTIONS } from '../../queries/index';
 import { useQuery } from '@apollo/react-hooks';
 import BeatLoader from 'react-spinners/BeatLoader';
-import DrpBtn from '../components/dropButton';
-import { AtoZ, Recent, ContactList } from '../components/contactList';
+import DrpBtn from '../../components/dropButton';
+// import { AtoZ, Recent, ContactList } from '../../components/contactList';
 
 // import components
-import ErrorPage from './errorpage';
+import ErrorPage from '../errorpage';
 
 const Contacts = ({ navigate }) => {
   let { loading, error, data } = useQuery(GET_USER_CONNECTIONS);
